+++ docs/changelog.md
@@ -1,3 +1,5 @@
+- Add the validator version to the validation OperationOutcome, as per the FHIR Tooling Extensions (#581)
+
 2026/09/14 Release 4.1.16
 
 - Fix starting the MCP server in onlyOneEngine mode